Introduce a new BinaryLogger logger. (#1747)
Introduce a new BinaryLogger.

/binaryLogger[:output.binlog]

Serializes all build events to a compressed binary file. By default the file is in the current directory and named "msbuild.binlog". The binary log is a detailed description of the build process that can later be used to reconstruct text logs and used by other analysis tools. A binary log is usually 10-20x smaller than the most detailed text diagnostic-level log, but it contains more information. (Short form: /bl)

Examples:
  /bl
  bl:output.binlog
  bl:..\..\custom.binlog
  binaryLogger

For now this commit only includes the logger (writer), but the reader is disconnected. In subsequent PRs we will add an option to playback the binary log into any other set of existing loggers.

Support for the new format will be added to MSBuild Structured Log Viewer so that you can view the .binlog files.

BuildEventArgsWriter and BuildEventArgsReader are general purpose custom binary serializers for BuildEventArgs objects.

BinaryLogReplayEventSource is a way to playback events from the log to subscribers of IEventSource.

A set of unit-tests is to follow in a separate PR.