Add a set of unit-test for serializing and deserializing BuildEventArgs.
Two minor fixes from code review feedback (Rainer).

The way the tests are written is objects are manually created, then serialized and deserialized back. The length of the bytes written needs to be the same as the number of bytes read back.

For each object type, a set of fields are compared on the original object and deserialized copy and we make sure the values are the same.