Versioned check for GetApiContractReferences (#1802)
If failed to get platform manifest, use the old way to return extension sdk references.

Correct paths for reference assemblies for systemManagementContract, part of the IOT extension SDK, are not resolved with insider preview SDK, 15003 and 15021. The files are in here:
C:\Program Files (x86)\Windows Kits\10\References<VERSION>\Windows.System.SystemManagementContract\3.0.0.0\Windows.System.SystemManagementContract.winmd.
The reason is when we call GetApiContractReferences we didn't check the Versioned property of the sdk. So it returned the path without