Log task inputs when using binary logger.
I've discovered that task inputs are not logged when using binary logger. This is because the logic here only logs task inputs if verbosity is Diagnostic or the env. variable MSBUILDLOGTASKINPUTS is set to 1.

http://source.dot.net/#MSBuild/XMake.cs,948
http://source.dot.net/#Microsoft.Build/BackEnd/TaskExecutionHost/TaskExecutionHost.cs,168

It is safe to set the "global" verbosity to diagnostic where I'm setting it, because from eyeballing the data flow from that point the only place it's consumed at is the check for log task inputs.