Don't set console color from FileLogger (#1827)
Because FileLogger used the default value of the SetColor and ResetColor
delegates, it was quietly changing and resetting console color for every
log message that was considered by the FileLogger. On Windows, this was
unnecessary but not too noticeable. On Unix systems, though, console
colors are set by emitting an escape code to the output stream, so this
became observable.

Explicitly passing the Dont* mechanisms to prevent the unnecessary work
from happening. I removed the null check for the `write` delegate from
ConsoleLogger's constructor because it seemed cleaner to set the
delegate to `null` from FileLogger than to explicitly specify the
default `Console.Out.Write`.

Fixes #1792.