Attempt to fix a race condition in TestBinaryLogRoundtrip.
My hypothesis is that when serializing properties of a ProjectStartedEventArgs we were iterating over the properties twice, first to establish count and the second time to actually write the items.

By inspecting the binary stream I've found that the log expected to find 166 property name/value pairs, but in the stream there's actually a 167th:

"SBuildLastTaskResult\u0004true\u0001\bTestItem\u0004Test\0\v5#Building with tools version \"15.0"

So I'm guessing is that the MSBuildLastTaskResult property is being set to true right after we grab the property count, but before we iterate over the properties. At least that would explain the extra property in the stream that we don't expect.

In any case the change seems good and I'd like to get this in (I'm 60% positive this will fix it).