Fix issue with /warnaserror on multi-proc builds (#1864)
When specifying `/warnaserror /m`, builds are expected to fail when warnings were treated as errors.  With MSBuild 15.0 RTM they were not failing.

This only repro'd on an optimized, ngen'd build out of VS which is why we didn't catch it. There was a timing issue where the overall build result was set too early before logging was complete and errors were still in the queue. Moving around the decision logic has fixed it. I tested this by doing a full build of Visual Studio and installing it on a test machine.

Closes #1791