Remove ProjectConfigurationsInferredFromUsage capability from the C#/VB targets. (#1879)
The legacy project system works without it (since it's not CPS based) and there are changes in the pipeline to use a different capability to defined configurations differently. On the new project system the capability is defined by the project system (as a default). Additionally the capability was also defined in the design time targets so there is some redundancy. Having the old capabilities here conflicts with that and requires a workaround to add a remove statement elsewhere in the targets.