Allow StreamReader to detect file Encoding w/o BOM. (#1904)
Passing in UTF8 w/o BOM Encoding object to StreamReader will allow the
StreamReader class to correctly detect encoding when a BOM exists in
the file. Using the default (UTF8 with BOM) will cause issues when the
file is UTF8 encoded without a BOM (added on save).

The overall issue was fixed in 8e3adba by reading the Stream. This
change allows for that behavior without needing to check the preamble
again.