Fix stackoverflow in preprocessor (#1909)
When updating the pre-processor for the implicit imports, I changed the internal map from an XmlElement to a string.  However, the outer XML of imports can be duplicated because it can contain a property or a relative path.  This map could contain duplicates and the pre-processor would process the same import over and over until a stack overflow occurred.

This fix restores the mapping of XmlElements and adds the implicit imports to the mapping after their nodes are created.

I also added a unit test for this case.

Fixes #1893