Enable RID-specific Apps on Shared Framework
Allow .NET Core applications to be RID-specific, but still run on the Shared Framework.  This enables an application to carry only the native assets it needs for the specific runtime it intends to run on.

For example, if an application uses a graphics library and wants to run on the shared framework, it would carry that graphics library for all the runtimes the graphics library supported (win, osx, linux, x86, x64, arm, etc).  But if the application is only intended to run on one runtime, these extra assemblies bloat the app.