Fix passing /p:DISABLE_CROSSGEN on cli.
When passing a property using the commandline, it is not flown through to "dotnet publish" calls.  So instead of using "dotnet publish", just use the MSBuild task to invoke the Publish target on the reidst project.  This way properties flow correctly.

I also did a little cleaning of unused build logic.