Search for types by full namespace
Some out-of-proc tests were failing on .NET Core because `System.Diagnostics.Process` is in its own assembly in .NET Core, but `GetTypeFromAssemblyUsingNamespace` always truncated the last element of the type name before searching for an assembly.

Fixing that by starting with the full specified type name and then continuing down the lop-off-last-element-and-try-again path.