Add support for SDK Resolvers * Adds support for MSBuild to discover plug-in based resolvers for the 'SDK' import feature. A resolver assembly can be placed in:     MSBuild\SdkResolver\(ResolverName)\(ResolverName).dll
MSBuild will create an instance of any class that implements SdkResolver
and attempt to resolve all SDK references with those resolver (sorted by
SdkResolver.Priority).

* Adds support for MSBuild to consume an Sdk reference as element (as an
alternative to a property on the Project element). Both versions of the
syntax will add an implicit Import element at the top and the bottom of
the file.

Related to #1493