Add a `baseDirectory` argument for LooksLikeUnixFilePath and ..
.. MaybeAdjustFilePath. The former checks if the first segment of the
path might be the name of an existent directory. This depends on the
CWD, but in some cases we might need to check this in a different
directory.

For example, project A initiates a build of project B, and while B is
being evaluated the CWD may not have been updated, but the paths in B's
project file are relative to B. So, the checks should be done in B's
directory.

This will be useful in an upcoming fix.