Fix file path in item metadata, when CWD != project directory
If a project A has:

    <Reference Include="foo">
        <HintPath>lib\foo.dll</HintPath>
    </Reference>

.. we want to convert the HintPath metadata to unix style. And that
happens in Expander.ExpandInto* methods.

But if the solution is building A, then at evaluation time the CWD may not
be set to the project A's directory, which would cause the
`LooksLikeUnixFilePath` test to fail for `lib\foo.dll`, since `lib` does
not exist in the solution's directory. And Expander methods don't have
the context to know about this.

So, instead we do this conversion in LazyItemOperation and attempt to
convert the path, as if we were in that project A's directory.

    #/abc/foo/foo.sln
    abc/ $ msbuild foo/foo.sln

Or basically any case where the CWD doesn't match project A's directory.

The test project fails in ResolveAssemblyReference with:

warning MSB3246: Resolved file has a bad image, no metadata, or is otherwise inaccessible. Could not find file "/Users/ankit/Projects/tests/nuget-tests/ConsoleApp2/ConsoleApp2/lib\fsharpFailsToEmbedResources.exe"
.. or ..
warning MSB3245: Could not resolve this reference. Could not locate the assembly "fsharpFailsToEmbedResources, Version=1.0.0.0". Check to make sure the assembly exists on disk. If this reference is required by your code, you may get compilation errors.

Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=55629