Use AssemblyName.Clone when possible. (#2016)
AssemblyName.Clone() is not available in early versions of .NET Core.
Until we can upgrade MSBuild to use .NET Core 2.0, we can't call the
Clone method directly. This change will attempt to call .Clone using
reflection and if not available fallback to the previous .NET Core
behavior.

Potential fix for #2015 (when running on .NET Core 2.0).