Project instance serialization (#2018)
Resolves #1726 

From issue comments:

Previous to this change, when a build is requested on a ProjectInstance and an out of proc node is used, the out of proc node uses two ProjectInstance objects to recover the state. ProjectInstance1 is the deserialized one. Apparently serializing a ProjectInstance only keeps part of the state: global properties, properties, and items. Everything else (targets, using tasks, etc) is discarded. To recover the missing part of the state, the out of proc node Re-evaluates the entry project form disk into ProjectInstance2. It then merges the two by keeping the properties and items from ProjectInstance1 and everything else from ProjectInstance2.

If the properties and items from ProjectInstance1 influence what targets get loaded, and what imports get imported, and if the project files change while the build is running in a way that changes those properties and items, then weird stuff like this failure happens (target not found exception).

The fix is to change ProjectInstance to serialize the complete evaluated state, and the out of proc node to reuse it entirely without re-evaluation from disk.