Fix silent test failure in BuildManager_Tests Addresses an issue in VerifyMultipleRequestForSameProjectWithErrors_ErrorAndContinue where the test will fail (writing an MSBuild_*.failure.txt file in the temp folder) but is not detected because a failure was expected. The primary change is to detect the number of MSBuild_*.failure.txt files and assert the value is unchanged before and after the test runs.
I used this opportunity to write TestEnvironment helper class to track
environmental variants before and after an individual test executes.
This class defines a common set of things that should not change (temp
path, environment variables, etc.) as a TestVariant and things that are
expected to change (environment variables, temp files/folders, etc.) as
a TestInvariant which is automatically reverted when the TestEnvironment
is disposed. This class is generic and can be extended to additional
cases.

Related to #1808