Avoid errors in conditions when MSBuild is evaluating for design time
When evaluating for design time, MSBuild will evaluate items even when they have a condition that is false.  So in a cross-targeting
build, even though the target framework identifier won't be set, it will still evaluate the items in these item groups.  So we need
to have a defense against trying to do a version comparison against an empty version in these conditions themselves.