Do not crash worker nodes on shutdown+pipe close
The `ShutdownNodesAfterParallelBuild` test reliably reported node crashes
when everything seemed to be going fine. This turned out to be because `
`NodeProviderOutOfProcBase.ShutdownAllNodes()` sent a shutdown-request
packet and then *immediately* closed the communications pipe, so
(depending on process scheduling) the pipe might not exist for the worker
node to send back its "I am shutting down" packet.

Fixes #2057 by swallowing exceptions when sending the `NodeShutdown`
packet.