Provide a way to not include the SDK Resolver
When building the entire stack for a non windows target, the windows
resolver packages will not be present (and in the case of a windows
build, you'd only have one architecture anyway). Provide a way to opt
out of including the resolver. It's only interesting for Visual
Studio insertion cases, which is something the community at large
doesn't need to worry about.