Embed the project and target source files from a build.
As the build is happening the BuildEventArgs will contain paths to the source files on disk encountered during the build: .proj files, .targets files, imported projects as well as potentially some source files.

It is valuable to capture these exact files and embed them into the binary log such that the log analysis has the actual sources using during the build at its disposal.

There are three options: don't capture sources, capture to a standalone .buildsources.zip archive and embed the zip archive directly into the binary log. The last option (embed) is the default. Logger parameters to change the default will be provided in a separate commit.

This commit adds the facility to write and read blobs to the log file, a SourceFileCollector that adds files to a .zip archive in the background and the integration pieces.