Don't convert backslashes that escape quotes in ToolTask (#2114)
Fixes #137 by converting only backslashes _not followed by a quote_ to
forward slashes when normalizing slashes in ToolTask.

Also makes it possible for a ToolTask to override this behavior to be
more correct for its purposes if necessary. A ToolTask that is known
to be cross-platform friendly can override 
AdjustCommandsForOperatingSystem to do nothing.