Fix issue where targets are executed after an error. (#2133)
Fixes an issue which occurs when two targets are scheduled to be built 
(e.g. ProduceError1 and ProduceError2), the condition of the first target
prevents it from executing, and a dependent target
(BeforeTargets='ProduceError1') fails. This change prevents
ProduceError2 from executing by marking the parent of the failure
(ProduceError1) with the Stop WorkUnitActionCode.

Fixes #2116