Find ref asms after ResolveReferences
e3a3d69 added a new FindReferenceAssembliesForReferences target, which
ran just before ResolveReferences. Dogfooding this revealed targets in
the VS SDK (at least) that manipulate @(ReferencePath) and are hooked in
as AfterTargets of ResolveReferences. That was silently ignored,
corrupting the user's build.

Fixes #2134 by moving FindReferenceAssembliesForReferences later in the
process, so anyone hooking ResolveReferences will see the changes they
make in the compiler.

This is not a fully general solution to the problem that someone might
hook an arbitrary point and manipulate @(ReferencePath), but it unbreaks
an obvious hook point.