Embed projects and imports in binary log
To improve the diagnosability of build logs a new option is added to collect project files and all imported projects used during the build.

There are three options: don't capture sources, capture to a standalone .ProjectImports.zip archive and embed the zip archive directly into the binary log. The last option (embed) is the default.