Fix Reference and PackageReference name collision
When building a DependencyContext, it is possible for
Reference names and PackageReference names to collide. We need
to ensure the DependencyContext libraries that are created have
unique names.

Also, when PreserveCompilationContext is false, we shouldn't be
creating Dependency objects for referenceassembly libraries, since
they won't be in the context.  They are only used for compilation.

Fix #1289