Simplify scheduling of ConflictResolution/NETStandard injection
SDK projects were running conflict-resolution too early, before ResolveProjectReferences
were resolved.  Fix this by moving conflict resolution out of
ResolvePackageDependenciesForBuildDependsOn and moving it later to only
a dependency of ResolveAssemblyReferences.

ImplicitlyExpandNETStandardFacades was previously trying to run as early as possible
which introduced problems in differing build sequences and complexity around handling
those.  Instead, run this as late as possible.  This is better because this target really
needs a full picture of all references to do its job.  I've scheduled it to run before
_HandlePackageFileConflicts and ResolveAssemblyReferences, just in case someone disables
_HandlePackageFileConflicts.

I've also renamed the extension conflict resolution target _HandlePackageFileConflicts
to match the SDK and made sure that it will never be brought into an SDK project.
This helped in the simplification of scheduling of ImplicitlyExpandNETStandardFacades.