Don't force Verbosity to Diag with /bl
It turns out that some tasks log a ridiculous amount of information as part of log task inputs. We see upwards of 5GB of strings logged as task inputs to Nuget.

To alleviate this pressure let's remove this, since the users will always have the easy option to manually add /v:diag to force logging task inputs.

For now not having this option is better since otherwise binary logger can crash people's builds.