@(TargetPathWithTargetPlatformMoniker) outside GetTargetPath
Some projects (and some NuGet packages) override GetTargetPath, and they
can't be expected to keep up with changes in the metadata now returned
by it.

This change moves the computation of the item returned from GetTargetPath
and Build into its own target (recycling
GetTargetPathWithTargetPlatformMoniker), so that it will still be
created by a project that overwrites GetTargetPath.

Fixes #2183.