Fix sequencing of _HandlePackageFileConflicts
Our previous sequencing assumed a later import (as was the case when
this came from a nuget package) and also assumed that nuget's targets
weren't necessarily imported.

To work with an earlier import we cannot make use of properties like
ResolveNuGetPackages and ProjectLockFile, but since we can depend on
NuGet targets being present we can schedule after ResolveNuGetPackageAssets
without causing log spew about missing targets.

As a result we can schedule ourselves Before RAR and After ResolveNuGetPackageAssets.
This ensures we'll run before RAR and as close to ResolveNuGetPackageAssets as
possible (if it runs).