Revert RAR AssemblyInformation caching
In .NET Core, it's safe to use an AssemblyInformation object without
disposing it, but that's not true in full framework because of the use
of COM objects to extract information from the assembly. Backing out the
cache until we can test a fix that correctly disposes the objects
(including for RAR failures and exceptions).

This reverts commit be56bc33a441ec97d91861a44d63c52a489e747f.

This reverts commit 55f846e7c290828f798fcba4c148708ad2d4510b.