Create dependencies for NETStandard.Library facades
The NETStandard.Library package (starting from version 2.0) adds
references to most assemblies via a .targets file rather than going
through the normal process to resolve assemblies from packages. This
causes the assemblies to show up under the Assemblies node in Solution
Explorer, rather than under the NETStandard.Library package where they
logically belong.

To surface these assemblies under the package we need to create new
`FileDependency` items linking them to the package. This is done in
the `PreprocessPackageDependenciesDesignTime` task as we don't want
to modify the normal build in any way, just surface some additional
information in the design-time build.

Note that this change does nothing to hide the items under the
Assemblies node; that will need to occur as a separate change in the
project system.