Hide import event logging under escape hatch
The logging of import events introduced a 100ms regression in c++
scenarios. Hiding them under an escape hatch for now. The binary logger
is the only thing that sets them on.

Later on, we should add another enum value to ProjectLoadSettings to
enable API users to run this. Or figure our how to flow debugger
verbosity into the Evaluator so we can branch on it.