Hide reparented facade assemblies
When reparenting a `Reference` item under the related NuGet package, also hide Facade assemblies.

For example, in NETStandard.Library we only want the user to see netstandard.dll, not the 100+ other facade assemblies added by the package.

This change requires further work in the project system to respect the "Visible" metadata.