Exec batch file encoding on Windows/core
Prior to this change, .NET Core MSBuild on Windows behaved like .NET
Core MSBuild on other OSes and treated its output as UTF-8. But that's
not true on Windows because of OS limitations, not .NET ones. This
commit brings back the logic used for full framework MSBuild (except the
unavailable is-Windows-too-old check).