Add ability to use newline separators in response files and update WinMDExp task to use it (#2216)
* Add ability to use new line separators in response files

Add new property to CommandLineBuilder and CommandLineBuilderExtension class to use newlines instead of spaces

Add new property to ToolTaskExtension to use newlines in response files instead of spaces.

* Update WinMDExp task to use a response file with newline separators and unquoted arguments

WinMDExp fails if there are quotes anywhere in the response file passed
to it, but correctly understands spaces in paths passed to /reference:
when the arguments are newline separated. Avoid ToolTask's default
quoting rules when populating the response file.